+++ src/remote.h
@@ -33,6 +33,7 @@ struct git_remote {
 	unsigned int need_pack;
 	git_remote_autotag_option_t download_tags;
 	int update_fetchhead;
+	int passed_refspecs;
 };
 
 const char* git_remote__urlfordirection(struct git_remote *remote, int direction);